Add bad request response for update and delete of non-existent users, instead of 404 response previously

+++ controllers/users.js
@@ -1,11 +1,12 @@
 const { hashPasword } = require('../utils/auth.js');
 const { db } = require('../config/db');
 const asyncHandler = require('../middleware/async');
+const ErrorResponse = require('../utils/errorResponse.js');
 
 /**
  * @desc    Get all users
  * @route   GET /api/users
- * @access   Private/Admin
+ * @access   Admin
  */
 exports.getUsers = asyncHandler(async (req, res) => {
   // TODO: Handle hiding of user credential, whilst allowing select queries
@@ -15,7 +16,7 @@ exports.getUsers = asyncHandler(async (req, res) => {
 /**
  * @desc    Get single user
  * @route   GET /api/users/:id
- * @access  Private/Admin
+ * @access  Admin
  */
 exports.getUser = asyncHandler(async (req, res) => {
   const rows = await db.one(
@@ -28,9 +29,9 @@ exports.getUser = asyncHandler(async (req, res) => {
 });
 
 /**
- * @desc    Create user and user profile
+ * @desc    Create user and associated user profile
  * @route   POST /api/users
- * @access  Private/Admin
+ * @access  Admin
  */
 exports.createUser = asyncHandler(async (req, res) => {
   const { name, email, password } = req.body;
@@ -38,6 +39,7 @@ exports.createUser = asyncHandler(async (req, res) => {
   const hashedPassword = await hashPasword(password);
 
   /**
+   * SQL Transaction, creating user and associated user profile
    * Returns an array of 2 json:
    * 1st json: User auth
    * 2nd json: User profile
@@ -61,9 +63,19 @@ exports.createUser = asyncHandler(async (req, res) => {
 /**
  * @desc    Update single user
  * @route   PUT /api/users/:id
- * @access  Private/Admin
+ * @access  Admin
  */
-exports.updateUser = asyncHandler(async (req, res) => {
+exports.updateUser = asyncHandler(async (req, res, next) => {
+  // check if user exists
+  const isValidUser = await db.oneOrNone(
+    `SELECT * FROM users WHERE user_id = ${req.params.id}`
+  );
+
+  // return bad request response if invalid user
+  if (!isValidUser) {
+    return next(new ErrorResponse(`User does not exist`, 400));
+  }
+
   const { name, email, password } = req.body;
   let updateUserQuery = `UPDATE users SET `;
   if (name) {
@@ -92,9 +104,19 @@ exports.updateUser = asyncHandler(async (req, res) => {
 /**
  * @desc    Delete single user
  * @route   DELETE /api/users/:id
- * @access  Private/Admin
+ * @access  Admin
  */
 exports.deleteUser = asyncHandler(async (req, res) => {
+  // check if user exists
+  const isValidUser = await db.oneOrNone(
+    `SELECT * FROM users WHERE user_id = ${req.params.id}`
+  );
+
+  // return bad request response if invalid user
+  if (!isValidUser) {
+    return next(new ErrorResponse(`User does not exist`, 400));
+  }
+
   const row = await db.one(
     `DELETE FROM users WHERE user_id = ${req.params.id} RETURNING *`
   );
